Rename package user_printer => userprint
Be more inline with
https://golang.org/doc/effective_go.html#package-names
